refactor geting the service at runtime
